d/control: provide x-terminal-emulator
Foot is Wayland-only, but in my perspective an "X terminal emulator" is
simply a graphical terminal emulator; I believe that the Policy refers
to X specifically because it has been _the_ synonym of graphical for
decades. Hence why I'm submitting this patch.

Policy states that terminal emulator alternatives _should_ have a
priority of 20, and the main foot binary does. At the same time,
footclient is installed as an additional alternative with lower
priority, 10, allowing users running a foot server in the background to
set the lightweight client as their default terminal emulator. As the
foot client depends on a running server, and no such server runs by
default, it has a lower priority so that it won't ever be automatically
selected by the alternatives system.

Gbp-Dch: Full